03_gateway/requirements.txt
# KithLy Global Protocol - Python Gateway Dependencies

# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Database
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.0

# HTTP Client
httpx>=0.26.0

# Google AI / Gemini
google-generativeai>=0.3.0

# Notifications
firebase-admin>=6.3.0
twilio>=8.10.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
black>=24.1.0
ruff>=0.1.0
